requirements.txt
streamlit>=1.30
pytest>=7.0
python-dotenv>=1.0
numpy>=1.20
google-generativeai>=0.8
